@@ -43,7 +43,7 @@
         }
     ],
     "require": {
-        "php": "^8.1.0"
+        "php": "^8.2.0"
     },
     "require-dev": {
         "codedungeon/phpunit-result-printer": "^0.32.0",
