@@ -8,6 +8,7 @@ use ReflectionClass;
 use Rinvex\Country\Country;
 use PHPUnit\Framework\TestCase;
 use Rinvex\Country\CountryLoader;
+use PHPUnit\Framework\Attributes\Test;
 use Rinvex\Country\CountryLoaderException;
 
 class CountryLoaderTest extends TestCase
@@ -17,12 +18,12 @@ class CountryLoaderTest extends TestCase
 
     public static function setUpBeforeClass(): void
     {
-        $reflectedLoader = new ReflectionClass(CountryLoader::class);
-        self::$methods['get'] = $reflectedLoader->getMethod('get');
-        self::$methods['pluck'] = $reflectedLoader->getMethod('pluck');
-        self::$methods['filter'] = $reflectedLoader->getMethod('filter');
-        self::$methods['getFile'] = $reflectedLoader->getMethod('getFile');
-        self::$methods['collapse'] = $reflectedLoader->getMethod('collapse');
+        $reflectionClass = new ReflectionClass(CountryLoader::class);
+        self::$methods['get'] = $reflectionClass->getMethod('get');
+        self::$methods['pluck'] = $reflectionClass->getMethod('pluck');
+        self::$methods['filter'] = $reflectionClass->getMethod('filter');
+        self::$methods['getFile'] = $reflectionClass->getMethod('getFile');
+        self::$methods['collapse'] = $reflectionClass->getMethod('collapse');
 
         foreach (self::$methods as $method) {
             $method->setAccessible(true);
@@ -34,8 +35,8 @@ class CountryLoaderTest extends TestCase
         self::$methods = null;
     }
 
-    /** @test */
-    public function it_returns_country_data()
+    #[Test]
+    public function it_returns_country_data(): void
     {
         $countryArray = [
             'name' => [
@@ -140,8 +141,8 @@ class CountryLoaderTest extends TestCase
         $this->assertEquals(new Country($countryArray), CountryLoader::country('eg'));
     }
 
-    /** @test */
-    public function it_gets_data_with_where_conditions()
+    #[Test]
+    public function it_gets_data_with_where_conditions(): void
     {
         $this->assertEquals(['as', 'au', 'ck', 'fj', 'fm', 'gu', 'ki', 'mh', 'mp', 'nc', 'nf', 'nu', 'nr', 'nz', 'pn', 'pw', 'pg', 'pf', 'sb', 'tk', 'tl', 'to', 'tv', 'um', 'vu', 'wf', 'ws'], array_keys(CountryLoader::where('geo.continent', ['OC' => 'Oceania'])));
         $this->assertEquals('Egypt', current(CountryLoader::where('capital', '=', 'Cairo'))['name']['common']);
@@ -157,8 +158,8 @@ class CountryLoaderTest extends TestCase
         $this->assertEquals(19, count(array_keys(CountryLoader::where('dialling.national_number_lengths.0', '<', 5))));
     }
 
-    /** @test */
-    public function it_returns_country_array_shortlist()
+    #[Test]
+    public function it_returns_country_array_shortlist(): void
     {
         $this->assertEquals(250, count(CountryLoader::countries()));
         $this->assertIsArray(CountryLoader::countries()['eg']);
@@ -166,8 +167,8 @@ class CountryLoaderTest extends TestCase
         $this->assertArrayNotHasKey('geo', CountryLoader::countries()['eg']);
     }
 
-    /** @test */
-    public function it_returns_country_hydrated_shortlist()
+    #[Test]
+    public function it_returns_country_hydrated_shortlist(): void
     {
         $this->assertEquals(250, count(CountryLoader::countries(false, true)));
         $this->assertIsObject(CountryLoader::countries(false, true)['eg']);
@@ -175,8 +176,8 @@ class CountryLoaderTest extends TestCase
         $this->assertNull(CountryLoader::countries(false, true)['eg']->getGeodata());
     }
 
-    /** @test */
-    public function it_returns_country_array_longlist()
+    #[Test]
+    public function it_returns_country_array_longlist(): void
     {
         $this->assertEquals(250, count(CountryLoader::countries(true)));
         $this->assertIsArray(CountryLoader::countries(true)['eg']);
@@ -185,8 +186,8 @@ class CountryLoaderTest extends TestCase
         $this->assertArrayHasKey('geo', CountryLoader::countries(true)['eg']);
     }
 
-    /** @test */
-    public function it_returns_country_hydrated_longlist()
+    #[Test]
+    public function it_returns_country_hydrated_longlist(): void
     {
         $this->assertEquals(250, count(CountryLoader::countries(true, true)));
         $this->assertIsObject(CountryLoader::countries(true, true)['eg']);
@@ -195,33 +196,29 @@ class CountryLoaderTest extends TestCase
         $this->assertIsArray(CountryLoader::countries(true, true)['eg']->getGeodata());
     }
 
-    /** @test */
-    public function it_throws_an_exception_when_invalid_country()
+    #[Test]
+    public function it_throws_an_exception_when_invalid_country(): void
     {
         $this->expectException(CountryLoaderException::class);
 
         CountryLoader::country('asd');
     }
 
-    /** @test */
-    public function it_filters_data()
+    #[Test]
+    public function it_filters_data(): void
     {
         $array1 = [['id' => 1, 'name' => 'Hello'], ['id' => 2, 'name' => 'World']];
-        $this->assertEquals([1 => ['id' => 2, 'name' => 'World']], self::$methods['filter']->invoke(null, $array1, function ($item) {
-            return $item['id'] === 2;
-        }));
+        $this->assertEquals([1 => ['id' => 2, 'name' => 'World']], self::$methods['filter']->invoke(null, $array1, fn ($item): bool => $item['id'] === 2));
 
         $array2 = ['', 'Hello', '', 'World'];
         $this->assertEquals(['Hello', 'World'], array_values(self::$methods['filter']->invoke(null, $array2)));
 
         $array3 = ['id' => 1, 'first' => 'Hello', 'second' => 'World'];
-        $this->assertEquals(['first' => 'Hello', 'second' => 'World'], self::$methods['filter']->invoke(null, $array3, function ($item, $key) {
-            return $key !== 'id';
-        }));
+        $this->assertEquals(['first' => 'Hello', 'second' => 'World'], self::$methods['filter']->invoke(null, $array3, fn ($item, $key): bool => $key !== 'id'));
     }
 
-    /** @test */
-    public function it_gets_data()
+    #[Test]
+    public function it_gets_data(): void
     {
         $object = (object) ['users' => ['name' => ['Taylor', 'Otwell']]];
         $array = [(object) ['users' => [(object) ['name' => 'Taylor']]]];
@@ -230,22 +227,20 @@ class CountryLoaderTest extends TestCase
         $this->assertEquals('Taylor', self::$methods['get']->invoke(null, $array, '0.users.0.name'));
         $this->assertNull(self::$methods['get']->invoke(null, $array, '0.users.3'));
         $this->assertEquals('Not found', self::$methods['get']->invoke(null, $array, '0.users.3', 'Not found'));
-        $this->assertEquals('Not found', self::$methods['get']->invoke(null, $array, '0.users.3', function () {
-            return 'Not found';
-        }));
+        $this->assertEquals('Not found', self::$methods['get']->invoke(null, $array, '0.users.3', fn (): string => 'Not found'));
         $this->assertEquals('Taylor', self::$methods['get']->invoke(null, $dottedArray, ['users', 'first.name']));
         $this->assertNull(self::$methods['get']->invoke(null, $dottedArray, ['users', 'middle.name']));
         $this->assertEquals('Not found', self::$methods['get']->invoke(null, $dottedArray, ['users', 'last.name'], 'Not found'));
     }
 
-    /** @test */
-    public function it_returns_target_when_missing_key()
+    #[Test]
+    public function it_returns_target_when_missing_key(): void
     {
         $this->assertEquals(['test'], self::$methods['get']->invoke(null, ['test'], null));
     }
 
-    /** @test */
-    public function it_gets_data_with_nested_arrays()
+    #[Test]
+    public function it_gets_data_with_nested_arrays(): void
     {
         $array = [
             ['name' => 'taylor', 'email' => 'taylorotwell@gmail.com'],
@@ -268,8 +263,8 @@ class CountryLoaderTest extends TestCase
         $this->assertNull(self::$methods['get']->invoke(null, $array, 'posts.*.date'));
     }
 
-    /** @test */
-    public function it_gets_data_with_nested_double_nested_arrays_and_collapses_result()
+    #[Test]
+    public function it_gets_data_with_nested_double_nested_arrays_and_collapses_result(): void
     {
         $array = [
             'posts' => [
@@ -299,8 +294,8 @@ class CountryLoaderTest extends TestCase
         $this->assertEquals([], self::$methods['get']->invoke(null, $array, 'posts.*.users.*.name'));
     }
 
-    /** @test */
-    public function it_plucks_array()
+    #[Test]
+    public function it_plucks_array(): void
     {
         $data = [
             'post-1' => [
@@ -335,16 +330,16 @@ class CountryLoaderTest extends TestCase
         $this->assertEquals([null, null], self::$methods['pluck']->invoke(null, $data, 'foo.bar'));
     }
 
-    /** @test */
-    public function it_plucks_array_with_array_and_object_values()
+    #[Test]
+    public function it_plucks_array_with_array_and_object_values(): void
     {
         $array = [(object) ['name' => 'taylor', 'email' => 'foo'], ['name' => 'dayle', 'email' => 'bar']];
         $this->assertEquals(['taylor', 'dayle'], self::$methods['pluck']->invoke(null, $array, 'name'));
         $this->assertEquals(['taylor' => 'foo', 'dayle' => 'bar'], self::$methods['pluck']->invoke(null, $array, 'email', 'name'));
     }
 
-    /** @test */
-    public function it_plucks_array_with_nested_keys()
+    #[Test]
+    public function it_plucks_array_with_nested_keys(): void
     {
         $array = [['user' => ['taylor', 'otwell']], ['user' => ['dayle', 'rees']]];
         $this->assertEquals(['taylor', 'dayle'], self::$methods['pluck']->invoke(null, $array, 'user.0'));
@@ -353,8 +348,8 @@ class CountryLoaderTest extends TestCase
         $this->assertEquals(['taylor' => 'otwell', 'dayle' => 'rees'], self::$methods['pluck']->invoke(null, $array, ['user', 1], ['user', 0]));
     }
 
-    /** @test */
-    public function it_plucks_array_with_nested_arrays()
+    #[Test]
+    public function it_plucks_array_with_nested_arrays(): void
     {
         $array = [
             [
@@ -376,21 +371,21 @@ class CountryLoaderTest extends TestCase
         $this->assertEquals([['foo'], [null, null]], self::$methods['pluck']->invoke(null, $array, 'users.*.email'));
     }
 
-    /** @test */
-    public function it_collapses_array()
+    #[Test]
+    public function it_collapses_array(): void
     {
         $array = [[1], [2], [3], ['foo', 'bar'], ['baz', 'boom']];
         $this->assertEquals([1, 2, 3, 'foo', 'bar', 'baz', 'boom'], self::$methods['collapse']->invoke(null, $array));
     }
 
-    /** @test */
-    public function it_gets_file_content()
+    #[Test]
+    public function it_gets_file_content(): void
     {
         $this->assertStringEqualsFile(__DIR__.'/../../resources/data/eg.json', self::$methods['getFile']->invoke(null, __DIR__.'/../../resources/data/eg.json'));
     }
 
-    /** @test */
-    public function it_throws_an_exception_when_invalid_file()
+    #[Test]
+    public function it_throws_an_exception_when_invalid_file(): void
     {
         $this->expectException(CountryLoaderException::class);
 
