@@ -7,6 +7,7 @@ namespace Rinvex\Country\Tests\Unit;
 use Exception;
 use Rinvex\Country\Country;
 use PHPUnit\Framework\TestCase;
+use PHPUnit\Framework\Attributes\Test;
 
 class CountryTest extends TestCase
 {
@@ -16,10 +17,10 @@ class CountryTest extends TestCase
     /** @var array */
     protected $longAttributes;
 
-    /** @var \Rinvex\Country\Country */
+    /** @var Country */
     protected $shortCountry;
 
-    /** @var \Rinvex\Country\Country */
+    /** @var Country */
     protected $longCountry;
 
     protected function setUp(): void
@@ -161,16 +162,16 @@ class CountryTest extends TestCase
         $this->longCountry = new Country($this->longAttributes);
     }
 
-    /** @test */
-    public function it_throws_an_exception_when_missing_mandatory_attributes()
+    #[Test]
+    public function it_throws_an_exception_when_missing_mandatory_attributes(): void
     {
         $this->expectException(Exception::class);
 
         new Country([]);
     }
 
-    /** @test */
-    public function it_sets_attributes_once_instantiated()
+    #[Test]
+    public function it_sets_attributes_once_instantiated(): void
     {
         $this->assertEquals($this->shortAttributes['name'], $this->shortCountry->getName());
         $this->assertEquals($this->shortAttributes['official_name'], $this->shortCountry->getOfficialName());
@@ -180,1134 +181,1134 @@ class CountryTest extends TestCase
         $this->assertEquals('EGY', $this->shortCountry->getIsoAlpha3());
     }
 
-    /** @test */
-    public function it_gets_attributes()
+    #[Test]
+    public function it_gets_attributes(): void
     {
         $this->assertEquals($this->shortAttributes, $this->shortCountry->getAttributes());
     }
 
-    /** @test */
-    public function it_sets_attributes()
+    #[Test]
+    public function it_sets_attributes(): void
     {
         $this->shortCountry->setAttributes(['capital' => 'Cairo']);
 
         $this->assertEquals('Cairo', $this->shortCountry->getCapital());
     }
 
-    /** @test */
-    public function it_gets_dotted_attribute()
+    #[Test]
+    public function it_gets_dotted_attribute(): void
     {
         $this->assertEquals($this->shortAttributes['calling_code'], $this->shortCountry->get('calling_code'));
         $this->assertEquals($this->longAttributes['name']['native']['ara']['common'], $this->longCountry->get('name.native.ara.common'));
     }
 
-    /** @test */
-    public function it_gets_default_when_missing_value()
+    #[Test]
+    public function it_gets_default_when_missing_value(): void
     {
         $this->assertEquals('default', $this->shortCountry->get('unknown', 'default'));
     }
 
-    /** @test */
-    public function it_gets_all_attributes_when_missing_key()
+    #[Test]
+    public function it_gets_all_attributes_when_missing_key(): void
     {
         $this->assertEquals($this->shortAttributes, $this->shortCountry->get(null));
     }
 
-    /** @test */
-    public function it_sets_attribute()
+    #[Test]
+    public function it_sets_attribute(): void
     {
         $this->shortCountry->set('capital', 'Cairo');
 
         $this->assertEquals('Cairo', $this->shortCountry->getCapital());
     }
 
-    /** @test */
-    public function its_fluently_chainable_when_sets_attributes()
+    #[Test]
+    public function its_fluently_chainable_when_sets_attributes(): void
     {
         $this->assertEquals($this->shortCountry, $this->shortCountry->setAttributes([]));
     }
 
-    /** @test */
-    public function it_returns_name_from_longlist()
+    #[Test]
+    public function it_returns_name_from_longlist(): void
     {
         $this->assertEquals($this->longAttributes['name']['common'], $this->longCountry->getName());
     }
 
-    /** @test */
-    public function it_returns_name_from_shortlist()
+    #[Test]
+    public function it_returns_name_from_shortlist(): void
     {
         $this->assertEquals($this->shortAttributes['name'], $this->shortCountry->getName());
     }
 
-    /** @test */
-    public function it_returns_null_when_missing_name()
+    #[Test]
+    public function it_returns_null_when_missing_name(): void
     {
         $this->shortCountry->setAttributes([]);
 
         $this->assertNull($this->shortCountry->getName());
     }
 
-    /** @test */
-    public function it_returns_official_name_from_longlist()
+    #[Test]
+    public function it_returns_official_name_from_longlist(): void
     {
         $this->assertEquals($this->longAttributes['name']['official'], $this->longCountry->getOfficialName());
     }
 
-    /** @test */
-    public function it_returns_official_name_from_shortlist()
+    #[Test]
+    public function it_returns_official_name_from_shortlist(): void
     {
         $this->assertEquals($this->shortAttributes['official_name'], $this->shortCountry->getOfficialName());
     }
 
-    /** @test */
-    public function it_returns_null_when_missing_official_name()
+    #[Test]
+    public function it_returns_null_when_missing_official_name(): void
     {
         $this->shortCountry->setAttributes([]);
 
         $this->assertNull($this->shortCountry->getOfficialName());
     }
 
-    /** @test */
-    public function it_returns_native_name_from_longlist()
+    #[Test]
+    public function it_returns_native_name_from_longlist(): void
     {
         $this->assertEquals($this->longAttributes['name']['native']['ara']['common'], $this->longCountry->getNativeName());
     }
 
-    /** @test */
-    public function it_returns_native_name_from_shortlist()
+    #[Test]
+    public function it_returns_native_name_from_shortlist(): void
     {
         $this->assertEquals($this->shortAttributes['native_name'], $this->shortCountry->getNativeName());
     }
 
-    /** @test */
-    public function it_returns_null_when_missing_native_name()
+    #[Test]
+    public function it_returns_null_when_missing_native_name(): void
     {
         $this->shortCountry->setAttributes([]);
 
         $this->assertNull($this->shortCountry->getNativeName());
     }
 
-    /** @test */
-    public function it_returns_native_official_name_from_longlist()
+    #[Test]
+    public function it_returns_native_official_name_from_longlist(): void
     {
         $this->assertEquals($this->longAttributes['name']['native']['ara']['official'], $this->longCountry->getNativeOfficialName());
     }
 
-    /** @test */
-    public function it_returns_native_official_name_from_shortlist()
+    #[Test]
+    public function it_returns_native_official_name_from_shortlist(): void
     {
         $this->assertEquals($this->shortAttributes['native_official_name'], $this->shortCountry->getNativeOfficialName());
     }
 
-    /** @test */
-    public function it_returns_null_when_missing_native_official_name()
+    #[Test]
+    public function it_returns_null_when_missing_native_official_name(): void
     {
         $this->shortCountry->setAttributes([]);
 
         $this->assertNull($this->shortCountry->getNativeOfficialName());
     }
 
-    /** @test */
-    public function it_returns_array_of_native_names_from_longlist()
+    #[Test]
+    public function it_returns_array_of_native_names_from_longlist(): void
     {
         $this->assertNotEmpty($this->longCountry->getNativeNames());
         $this->assertEquals(current($this->longAttributes['name']['native']), current($this->longCountry->getNativeNames()));
     }
 
-    /** @test */
-    public function it_returns_null_when_missing_native_names()
+    #[Test]
+    public function it_returns_null_when_missing_native_names(): void
     {
         $this->shortCountry->setAttributes([]);
 
         $this->assertNull($this->shortCountry->getNativeNames());
     }
 
-    /** @test */
-    public function it_returns_demonym()
+    #[Test]
+    public function it_returns_demonym(): void
     {
         $this->assertEquals($this->longAttributes['demonym'], $this->longCountry->getDemonym());
     }
 
-    /** @test */
-    public function it_returns_null_when_missing_demonym()
+    #[Test]
+    public function it_returns_null_when_missing_demonym(): void
     {
         $this->longCountry->setAttributes([]);
 
         $this->assertNull($this->longCountry->getDemonym());
     }
 
-    /** @test */
-    public function it_returns_capital()
+    #[Test]
+    public function it_returns_capital(): void
     {
         $this->assertEquals($this->longAttributes['capital'], $this->longCountry->getCapital());
     }
 
-    /** @test */
-    public function it_returns_null_when_missing_capital()
+    #[Test]
+    public function it_returns_null_when_missing_capital(): void
     {
         $this->longCountry->setAttributes([]);
 
         $this->assertNull($this->longCountry->getCapital());
     }
 
-    /** @test */
-    public function it_returns_isoalpha2()
+    #[Test]
+    public function it_returns_isoalpha2(): void
     {
         $this->assertEquals($this->longAttributes['iso_3166_1_alpha2'], $this->longCountry->getIsoAlpha2());
     }
 
-    /** @test */
-    public function it_returns_null_when_missing_isoalpha2()
+    #[Test]
+    public function it_returns_null_when_missing_isoalpha2(): void
     {
         $this->longCountry->setAttributes([]);
 
         $this->assertNull($this->longCountry->getIsoAlpha2());
     }
 
-    /** @test */
-    public function it_returns_isoalpha3()
+    #[Test]
+    public function it_returns_isoalpha3(): void
     {
         $this->assertEquals($this->longAttributes['iso_3166_1_alpha3'], $this->longCountry->getIsoAlpha3());
     }
 
-    /** @test */
-    public function it_returns_null_when_missing_isoalpha3()
+    #[Test]
+    public function it_returns_null_when_missing_isoalpha3(): void
     {
         $this->longCountry->setAttributes([]);
 
         $this->assertNull($this->longCountry->getIsoAlpha3());
     }
 
-    /** @test */
-    public function it_returns_isonumeric()
+    #[Test]
+    public function it_returns_isonumeric(): void
     {
         $this->assertEquals($this->longAttributes['iso_3166_1_numeric'], $this->longCountry->getIsoNumeric());
     }
 
-    /** @test */
-    public function it_returns_null_when_missing_isonumeric()
+    #[Test]
+    public function it_returns_null_when_missing_isonumeric(): void
     {
         $this->longCountry->setAttributes([]);
 
         $this->assertNull($this->longCountry->getIsoNumeric());
     }
 
-    /** @test */
-    public function it_returns_currency()
+    #[Test]
+    public function it_returns_currency(): void
     {
         $this->assertEquals($this->longAttributes['currency']['EGP'], $this->longCountry->getCurrency());
     }
 
-    /** @test */
-    public function it_returns_first_currency_when_missing_requested_currency()
+    #[Test]
+    public function it_returns_first_currency_when_missing_requested_currency(): void
     {
         $this->assertEquals($this->longAttributes['currency']['EGP'], $this->longCountry->getCurrency('USD'));
     }
 
-    /** @test */
-    public function it_returns_null_when_missing_currency()
+    #[Test]
+    public function it_returns_null_when_missing_currency(): void
     {
         $this->longCountry->setAttributes([]);
 
         $this->assertNull($this->longCountry->getCurrency());
     }
 
-    /** @test */
-    public function it_returns_currencies()
+    #[Test]
+    public function it_returns_currencies(): void
     {
         $this->assertEquals($this->longAttributes['currency'], $this->longCountry->getCurrencies());
     }
 
-    /** @test */
-    public function it_returns_null_when_missing_currencies()
+    #[Test]
+    public function it_returns_null_when_missing_currencies(): void
     {
         $this->longCountry->setAttributes([]);
 
         $this->assertNull($this->longCountry->getCurrencies());
     }
 
-    /** @test */
-    public function it_returns_tld()
+    #[Test]
+    public function it_returns_tld(): void
     {
         $this->assertEquals(current($this->longAttributes['tld']), $this->longCountry->getTld());
     }
 
-    /** @test */
-    public function it_returns_null_when_missing_tld()
+    #[Test]
+    public function it_returns_null_when_missing_tld(): void
     {
         $this->longCountry->setAttributes([]);
 
         $this->assertNull($this->longCountry->getTld());
     }
 
-    /** @test */
-    public function it_returns_tlds()
+    #[Test]
+    public function it_returns_tlds(): void
     {
         $this->assertEquals($this->longAttributes['tld'], $this->longCountry->getTlds());
     }
 
-    /** @test */
-    public function it_returns_null_when_missing_tlds()
+    #[Test]
+    public function it_returns_null_when_missing_tlds(): void
     {
         $this->longCountry->setAttributes([]);
 
         $this->assertNull($this->longCountry->getTlds());
     }
 
-    /** @test */
-    public function it_returns_altspellings()
+    #[Test]
+    public function it_returns_altspellings(): void
     {
         $this->assertEquals($this->longAttributes['alt_spellings'], $this->longCountry->getAltSpellings());
     }
 
-    /** @test */
-    public function it_returns_null_when_missing_altspellings()
+    #[Test]
+    public function it_returns_null_when_missing_altspellings(): void
     {
         $this->longCountry->setAttributes([]);
 
         $this->assertNull($this->longCountry->getAltSpellings());
     }
 
-    /** @test */
-    public function it_returns_language()
+    #[Test]
+    public function it_returns_language(): void
     {
         $this->assertEquals($this->longAttributes['languages']['ara'], $this->longCountry->getLanguage());
     }
 
-    /** @test */
-    public function it_returns_first_currency_when_missing_requested_language()
+    #[Test]
+    public function it_returns_first_currency_when_missing_requested_language(): void
     {
         $this->assertEquals($this->longAttributes['languages']['ara'], $this->longCountry->getLanguage('eng'));
     }
 
-    /** @test */
-    public function it_returns_null_when_missing_language()
+    #[Test]
+    public function it_returns_null_when_missing_language(): void
     {
         $this->longCountry->setAttributes([]);
 
         $this->assertNull($this->longCountry->getLanguage());
     }
 
-    /** @test */
-    public function it_returns_languages()
+    #[Test]
+    public function it_returns_languages(): void
     {
         $this->assertEquals($this->longAttributes['languages'], $this->longCountry->getLanguages());
     }
 
-    /** @test */
-    public function it_returns_null_when_missing_languages()
+    #[Test]
+    public function it_returns_null_when_missing_languages(): void
     {
         $this->longCountry->setAttributes([]);
 
         $this->assertNull($this->longCountry->getLanguages());
     }
 
-    /** @test */
-    public function it_returns_translation()
+    #[Test]
+    public function it_returns_translation(): void
     {
         $this->assertEquals($this->longAttributes['name']['native']['ara'], $this->longCountry->getTranslation());
     }
 
-    /** @test */
-    public function it_returns_first_translation_when_missing_requested_translation()
+    #[Test]
+    public function it_returns_first_translation_when_missing_requested_translation(): void
     {
         $this->assertEquals($this->longAttributes['name']['native']['ara'], $this->longCountry->getTranslation('ara'));
     }
 
-    /** @test */
-    public function it_returns_translations()
+    #[Test]
+    public function it_returns_translations(): void
     {
         $this->assertEquals($this->longAttributes['name']['native']['ara'], $this->longCountry->getTranslations()['ara']);
     }
 
-    /** @test */
-    public function it_returns_first_translation_when_missing_requested_translations()
+    #[Test]
+    public function it_returns_first_translation_when_missing_requested_translations(): void
     {
         $this->assertEquals($this->longAttributes['name']['native']['ara'], $this->longCountry->getTranslation('ara'));
     }
 
-    /** @test */
-    public function it_returns_geodata()
+    #[Test]
+    public function it_returns_geodata(): void
     {
         $this->assertEquals($this->longAttributes['geo'], $this->longCountry->getGeodata());
     }
 
-    /** @test */
-    public function it_returns_null_when_missing_geodata()
+    #[Test]
+    public function it_returns_null_when_missing_geodata(): void
     {
         $this->longCountry->setAttributes([]);
 
         $this->assertNull($this->longCountry->getGeodata());
     }
 
-    /** @test */
-    public function it_returns_continent()
+    #[Test]
+    public function it_returns_continent(): void
     {
         $this->assertEquals(current($this->longAttributes['geo']['continent']), $this->longCountry->getContinent());
     }
 
-    /** @test */
-    public function it_returns_null_when_missing_continent()
+    #[Test]
+    public function it_returns_null_when_missing_continent(): void
     {
         $this->longCountry->setAttributes([]);
 
         $this->assertNull($this->longCountry->getContinent());
     }
 
-    /** @test */
-    public function it_returns_postal_code()
+    #[Test]
+    public function it_returns_postal_code(): void
     {
         $this->assertEquals($this->longAttributes['geo']['postal_code'], $this->longCountry->usesPostalCode());
     }
 
-    /** @test */
-    public function it_returns_null_when_missing_postal_code()
+    #[Test]
+    public function it_returns_null_when_missing_postal_code(): void
     {
         $this->longCountry->setAttributes([]);
 
         $this->assertNull($this->longCountry->usesPostalCode());
     }
 
-    /** @test */
-    public function it_returns_latitude()
+    #[Test]
+    public function it_returns_latitude(): void
     {
         $this->assertEquals($this->longAttributes['geo']['latitude'], $this->longCountry->getLatitude());
     }
 
-    /** @test */
-    public function it_returns_null_when_missing_latitude()
+    #[Test]
+    public function it_returns_null_when_missing_latitude(): void
     {
         $this->longCountry->setAttributes([]);
 
         $this->assertNull($this->longCountry->getLatitude());
     }
 
-    /** @test */
-    public function it_returns_latitude_desc()
+    #[Test]
+    public function it_returns_latitude_desc(): void
     {
         $this->assertEquals($this->longAttributes['geo']['latitude_desc'], $this->longCountry->getLatitudeDesc());
     }
 
-    /** @test */
-    public function it_returns_null_when_missing_latitude_desc()
+    #[Test]
+    public function it_returns_null_when_missing_latitude_desc(): void
     {
         $this->longCountry->setAttributes([]);
 
         $this->assertNull($this->longCountry->getLatitudeDesc());
     }
 
-    /** @test */
-    public function it_returns_max_latitude()
+    #[Test]
+    public function it_returns_max_latitude(): void
     {
         $this->assertEquals($this->longAttributes['geo']['max_latitude'], $this->longCountry->getMaxLatitude());
     }
 
-    /** @test */
-    public function it_returns_null_when_missing_lmax_latitude()
+    #[Test]
+    public function it_returns_null_when_missing_lmax_latitude(): void
     {
         $this->longCountry->setAttributes([]);
 
         $this->assertNull($this->longCountry->getMaxLatitude());
     }
 
-    /** @test */
-    public function it_returns_longitude()
+    #[Test]
+    public function it_returns_longitude(): void
     {
         $this->assertEquals($this->longAttributes['geo']['longitude'], $this->longCountry->getLongitude());
     }
 
-    /** @test */
-    public function it_returns_null_when_missing_longitude()
+    #[Test]
+    public function it_returns_null_when_missing_longitude(): void
     {
         $this->longCountry->setAttributes([]);
 
         $this->assertNull($this->longCountry->getLongitude());
     }
 
-    /** @test */
-    public function it_returns_longitude_desc()
+    #[Test]
+    public function it_returns_longitude_desc(): void
     {
         $this->assertEquals($this->longAttributes['geo']['longitude_desc'], $this->longCountry->getLongitudeDesc());
     }
 
-    /** @test */
-    public function it_returns_null_when_missing_longitude_desc()
+    #[Test]
+    public function it_returns_null_when_missing_longitude_desc(): void
     {
         $this->longCountry->setAttributes([]);
 
         $this->assertNull($this->longCountry->getLongitudeDesc());
     }
 
-    /** @test */
-    public function it_returns_max_longitude()
+    #[Test]
+    public function it_returns_max_longitude(): void
     {
         $this->assertEquals($this->longAttributes['geo']['max_longitude'], $this->longCountry->getMaxLongitude());
     }
 
-    /** @test */
-    public function it_returns_null_when_missing_max_longitude()
+    #[Test]
+    public function it_returns_null_when_missing_max_longitude(): void
     {
         $this->longCountry->setAttributes([]);
 
         $this->assertNull($this->longCountry->getMaxLongitude());
     }
 
-    /** @test */
-    public function it_returns_min_longitude()
+    #[Test]
+    public function it_returns_min_longitude(): void
     {
         $this->assertEquals($this->longAttributes['geo']['min_longitude'], $this->longCountry->getMinLongitude());
     }
 
-    /** @test */
-    public function it_returns_null_when_missing_min_longitude()
+    #[Test]
+    public function it_returns_null_when_missing_min_longitude(): void
     {
         $this->longCountry->setAttributes([]);
 
         $this->assertNull($this->longCountry->getMinLongitude());
     }
 
-    /** @test */
-    public function it_returns_min_latitude()
+    #[Test]
+    public function it_returns_min_latitude(): void
     {
         $this->assertEquals($this->longAttributes['geo']['min_latitude'], $this->longCountry->getMinLatitude());
     }
 
-    /** @test */
-    public function it_returns_null_when_missing_min_latitude()
+    #[Test]
+    public function it_returns_null_when_missing_min_latitude(): void
     {
         $this->longCountry->setAttributes([]);
 
         $this->assertNull($this->longCountry->getMinLatitude());
     }
 
-    /** @test */
-    public function it_returns_area()
+    #[Test]
+    public function it_returns_area(): void
     {
         $this->assertEquals($this->longAttributes['geo']['area'], $this->longCountry->getArea());
     }
 
-    /** @test */
-    public function it_returns_null_when_missing_area()
+    #[Test]
+    public function it_returns_null_when_missing_area(): void
     {
         $this->longCountry->setAttributes([]);
 
         $this->assertNull($this->longCountry->getArea());
     }
 
-    /** @test */
-    public function it_returns_region()
+    #[Test]
+    public function it_returns_region(): void
     {
         $this->assertEquals($this->longAttributes['geo']['region'], $this->longCountry->getRegion());
     }
 
-    /** @test */
-    public function it_returns_null_when_missing_region()
+    #[Test]
+    public function it_returns_null_when_missing_region(): void
     {
         $this->longCountry->setAttributes([]);
 
         $this->assertNull($this->longCountry->getRegion());
     }
 
-    /** @test */
-    public function it_returns_subregion()
+    #[Test]
+    public function it_returns_subregion(): void
     {
         $this->assertEquals($this->longAttributes['geo']['subregion'], $this->longCountry->getSubregion());
     }
 
-    /** @test */
-    public function it_returns_null_when_missing_subregion()
+    #[Test]
+    public function it_returns_null_when_missing_subregion(): void
     {
         $this->longCountry->setAttributes([]);
 
         $this->assertNull($this->longCountry->getSubregion());
     }
 
-    /** @test */
-    public function it_returns_world_region()
+    #[Test]
+    public function it_returns_world_region(): void
     {
         $this->assertEquals($this->longAttributes['geo']['world_region'], $this->longCountry->getWorldRegion());
     }
 
-    /** @test */
-    public function it_returns_null_when_missing_world_region()
+    #[Test]
+    public function it_returns_null_when_missing_world_region(): void
     {
         $this->longCountry->setAttributes([]);
 
         $this->assertNull($this->longCountry->getWorldRegion());
     }
 
-    /** @test */
-    public function it_returns_region_code()
+    #[Test]
+    public function it_returns_region_code(): void
     {
         $this->assertEquals($this->longAttributes['geo']['region_code'], $this->longCountry->getRegionCode());
     }
 
-    /** @test */
-    public function it_returns_null_when_missing_region_code()
+    #[Test]
+    public function it_returns_null_when_missing_region_code(): void
     {
         $this->longCountry->setAttributes([]);
 
         $this->assertNull($this->longCountry->getRegionCode());
     }
 
-    /** @test */
-    public function it_returns_subregion_code()
+    #[Test]
+    public function it_returns_subregion_code(): void
     {
         $this->assertEquals($this->longAttributes['geo']['subregion_code'], $this->longCountry->getSubregionCode());
     }
 
-    /** @test */
-    public function it_returns_null_when_missing_subregion_code()
+    #[Test]
+    public function it_returns_null_when_missing_subregion_code(): void
     {
         $this->longCountry->setAttributes([]);
 
         $this->assertNull($this->longCountry->getSubregionCode());
     }
 
-    /** @test */
-    public function it_returns_landlocked_status()
+    #[Test]
+    public function it_returns_landlocked_status(): void
     {
         $this->assertEquals($this->longAttributes['geo']['landlocked'], $this->longCountry->isLandlocked());
     }
 
-    /** @test */
-    public function it_returns_null_when_missing_landlocked_status()
+    #[Test]
+    public function it_returns_null_when_missing_landlocked_status(): void
     {
         $this->longCountry->setAttributes([]);
 
         $this->assertNull($this->longCountry->isLandlocked());
     }
 
-    /** @test */
-    public function it_returns_borders()
+    #[Test]
+    public function it_returns_borders(): void
     {
         $this->assertEquals($this->longAttributes['geo']['borders'], $this->longCountry->getBorders());
     }
 
-    /** @test */
-    public function it_returns_null_when_missing_borders()
+    #[Test]
+    public function it_returns_null_when_missing_borders(): void
     {
         $this->longCountry->setAttributes([]);
 
         $this->assertNull($this->longCountry->getBorders());
     }
 
-    /** @test */
-    public function it_returns_independent_status()
+    #[Test]
+    public function it_returns_independent_status(): void
     {
         $this->assertEquals($this->longAttributes['geo']['independent'], $this->longCountry->isIndependent());
     }
 
-    /** @test */
-    public function it_returns_null_when_missing_independent_status()
+    #[Test]
+    public function it_returns_null_when_missing_independent_status(): void
     {
         $this->longCountry->setAttributes([]);
 
         $this->assertNull($this->longCountry->isIndependent());
     }
 
-    /** @test */
-    public function it_returns_calling_code_from_longlist()
+    #[Test]
+    public function it_returns_calling_code_from_longlist(): void
     {
         $this->assertEquals(current($this->longAttributes['dialling']['calling_code']), $this->longCountry->getCallingCode());
     }
 
-    /** @test */
-    public function it_returns_calling_code_from_shortlist()
+    #[Test]
+    public function it_returns_calling_code_from_shortlist(): void
     {
         $this->assertEquals(current($this->shortAttributes['calling_code']), $this->shortCountry->getCallingCode());
     }
 
-    /** @test */
-    public function it_returns_null_when_missing_calling_code()
+    #[Test]
+    public function it_returns_null_when_missing_calling_code(): void
     {
         $this->longCountry->setAttributes([]);
 
         $this->assertNull($this->longCountry->getCallingCode());
     }
 
-    /** @test */
-    public function it_returns_calling_codes()
+    #[Test]
+    public function it_returns_calling_codes(): void
     {
         $this->assertEquals($this->longAttributes['dialling']['calling_code'], $this->longCountry->getCallingCodes());
     }
 
-    /** @test */
-    public function it_returns_null_when_missing_calling_codes()
+    #[Test]
+    public function it_returns_null_when_missing_calling_codes(): void
     {
         $this->longCountry->setAttributes([]);
 
         $this->assertNull($this->longCountry->getCallingCodes());
     }
 
-    /** @test */
-    public function it_returns_national_prefix()
+    #[Test]
+    public function it_returns_national_prefix(): void
     {
         $this->assertEquals($this->longAttributes['dialling']['national_prefix'], $this->longCountry->getNationalPrefix());
     }
 
-    /** @test */
-    public function it_returns_null_when_missing_national_prefix()
+    #[Test]
+    public function it_returns_null_when_missing_national_prefix(): void
     {
         $this->longCountry->setAttributes([]);
 
         $this->assertNull($this->longCountry->getNationalPrefix());
     }
 
-    /** @test */
-    public function it_returns_national_number_length()
+    #[Test]
+    public function it_returns_national_number_length(): void
     {
         $this->assertEquals(current($this->longAttributes['dialling']['national_number_lengths']), $this->longCountry->getNationalNumberLength());
     }
 
-    /** @test */
-    public function it_returns_null_when_missing_national_number_length()
+    #[Test]
+    public function it_returns_null_when_missing_national_number_length(): void
     {
         $this->longCountry->setAttributes([]);
 
         $this->assertNull($this->longCountry->getNationalNumberLength());
     }
 
-    /** @test */
-    public function it_returns_national_number_lengths()
+    #[Test]
+    public function it_returns_national_number_lengths(): void
     {
         $this->assertEquals($this->longAttributes['dialling']['national_number_lengths'], $this->longCountry->getNationalNumberLengths());
     }
 
-    /** @test */
-    public function it_returns_null_when_missing_national_number_lengths()
+    #[Test]
+    public function it_returns_null_when_missing_national_number_lengths(): void
     {
         $this->longCountry->setAttributes([]);
 
         $this->assertNull($this->longCountry->getNationalNumberLengths());
     }
 
-    /** @test */
-    public function it_returns_national_destination_code_length()
+    #[Test]
+    public function it_returns_national_destination_code_length(): void
     {
         $this->assertEquals(current($this->longAttributes['dialling']['national_destination_code_lengths']), $this->longCountry->getNationalDestinationCodeLength());
     }
 
-    /** @test */
-    public function it_returns_null_when_missing_national_destination_code_length()
+    #[Test]
+    public function it_returns_null_when_missing_national_destination_code_length(): void
     {
         $this->longCountry->setAttributes([]);
 
         $this->assertNull($this->longCountry->getNationalDestinationCodeLength());
     }
 
-    /** @test */
-    public function it_returns_national_destination_code_lengths()
+    #[Test]
+    public function it_returns_national_destination_code_lengths(): void
     {
         $this->assertEquals($this->longAttributes['dialling']['national_destination_code_lengths'], $this->longCountry->getNationalDestinationCodeLengths());
     }
 
-    /** @test */
-    public function it_returns_null_when_missing_national_destination_code_lengths()
+    #[Test]
+    public function it_returns_null_when_missing_national_destination_code_lengths(): void
     {
         $this->longCountry->setAttributes([]);
 
         $this->assertNull($this->longCountry->getNationalDestinationCodeLengths());
     }
 
-    /** @test */
-    public function it_returns_international_prefix()
+    #[Test]
+    public function it_returns_international_prefix(): void
     {
         $this->assertEquals($this->longAttributes['dialling']['international_prefix'], $this->longCountry->getInternationalPrefix());
     }
 
-    /** @test */
-    public function it_returns_null_when_missing_international_prefix()
+    #[Test]
+    public function it_returns_null_when_missing_international_prefix(): void
     {
         $this->longCountry->setAttributes([]);
 
         $this->assertNull($this->longCountry->getInternationalPrefix());
     }
 
-    /** @test */
-    public function it_returns_extra()
+    #[Test]
+    public function it_returns_extra(): void
     {
         $this->assertEquals($this->longAttributes['extra'], $this->longCountry->getExtra());
     }
 
-    /** @test */
-    public function it_returns_null_when_missing_extra()
+    #[Test]
+    public function it_returns_null_when_missing_extra(): void
     {
         $this->longCountry->setAttributes([]);
 
         $this->assertNull($this->longCountry->getExtra());
     }
 
-    /** @test */
-    public function it_returns_geonameid()
+    #[Test]
+    public function it_returns_geonameid(): void
     {
         $this->assertEquals($this->longAttributes['extra']['geonameid'], $this->longCountry->getGeonameid());
     }
 
-    /** @test */
-    public function it_returns_null_when_missing_geonameid()
+    #[Test]
+    public function it_returns_null_when_missing_geonameid(): void
     {
         $this->longCountry->setAttributes([]);
 
         $this->assertNull($this->longCountry->getGeonameid());
     }
 
-    /** @test */
-    public function it_returns_edgar()
+    #[Test]
+    public function it_returns_edgar(): void
     {
         $this->assertEquals($this->longAttributes['extra']['edgar'], $this->longCountry->getEdgar());
     }
 
-    /** @test */
-    public function it_returns_null_when_missing_edgar()
+    #[Test]
+    public function it_returns_null_when_missing_edgar(): void
     {
         $this->longCountry->setAttributes([]);
 
         $this->assertNull($this->longCountry->getEdgar());
     }
 
-    /** @test */
-    public function it_returns_itu()
+    #[Test]
+    public function it_returns_itu(): void
     {
         $this->assertEquals($this->longAttributes['extra']['itu'], $this->longCountry->getItu());
     }
 
-    /** @test */
-    public function it_returns_null_when_missing_itu()
+    #[Test]
+    public function it_returns_null_when_missing_itu(): void
     {
         $this->longCountry->setAttributes([]);
 
         $this->assertNull($this->longCountry->getItu());
     }
 
-    /** @test */
-    public function it_returns_marc()
+    #[Test]
+    public function it_returns_marc(): void
     {
         $this->assertEquals($this->longAttributes['extra']['marc'], $this->longCountry->getMarc());
     }
 
-    /** @test */
-    public function it_returns_null_when_missing_marc()
+    #[Test]
+    public function it_returns_null_when_missing_marc(): void
     {
         $this->longCountry->setAttributes([]);
 
         $this->assertNull($this->longCountry->getMarc());
     }
 
-    /** @test */
-    public function it_returns_wmo()
+    #[Test]
+    public function it_returns_wmo(): void
     {
         $this->assertEquals($this->longAttributes['extra']['wmo'], $this->longCountry->getWmo());
     }
 
-    /** @test */
-    public function it_returns_null_when_missing_wmo()
+    #[Test]
+    public function it_returns_null_when_missing_wmo(): void
     {
         $this->longCountry->setAttributes([]);
 
         $this->assertNull($this->longCountry->getWmo());
     }
 
-    /** @test */
-    public function it_returns_ds()
+    #[Test]
+    public function it_returns_ds(): void
     {
         $this->assertEquals($this->longAttributes['extra']['ds'], $this->longCountry->getDs());
     }
 
-    /** @test */
-    public function it_returns_null_when_missing_ds()
+    #[Test]
+    public function it_returns_null_when_missing_ds(): void
     {
         $this->longCountry->setAttributes([]);
 
         $this->assertNull($this->longCountry->getDs());
     }
 
-    /** @test */
-    public function it_returns_fifa()
+    #[Test]
+    public function it_returns_fifa(): void
     {
         $this->assertEquals($this->longAttributes['extra']['fifa'], $this->longCountry->getFifa());
     }
 
-    /** @test */
-    public function it_returns_null_when_missing_fifa()
+    #[Test]
+    public function it_returns_null_when_missing_fifa(): void
     {
         $this->longCountry->setAttributes([]);
 
         $this->assertNull($this->longCountry->getFifa());
     }
 
-    /** @test */
-    public function it_returns_fips()
+    #[Test]
+    public function it_returns_fips(): void
     {
         $this->assertEquals($this->longAttributes['extra']['fips'], $this->longCountry->getFips());
     }
 
-    /** @test */
-    public function it_returns_null_when_missing_fips()
+    #[Test]
+    public function it_returns_null_when_missing_fips(): void
     {
         $this->longCountry->setAttributes([]);
 
         $this->assertNull($this->longCountry->getFips());
     }
 
-    /** @test */
-    public function it_returns_gaul()
+    #[Test]
+    public function it_returns_gaul(): void
     {
         $this->assertEquals($this->longAttributes['extra']['gaul'], $this->longCountry->getGaul());
     }
 
-    /** @test */
-    public function it_returns_null_when_missing_gaul()
+    #[Test]
+    public function it_returns_null_when_missing_gaul(): void
     {
         $this->longCountry->setAttributes([]);
 
         $this->assertNull($this->longCountry->getGaul());
     }
 
-    /** @test */
-    public function it_returns_ioc()
+    #[Test]
+    public function it_returns_ioc(): void
     {
         $this->assertEquals($this->longAttributes['extra']['ioc'], $this->longCountry->getIoc());
     }
 
-    /** @test */
-    public function it_returns_null_when_missing_ioc()
+    #[Test]
+    public function it_returns_null_when_missing_ioc(): void
     {
         $this->longCountry->setAttributes([]);
 
         $this->assertNull($this->longCountry->getIoc());
     }
 
-    /** @test */
-    public function it_returns_cowc()
+    #[Test]
+    public function it_returns_cowc(): void
     {
         $this->assertEquals($this->longAttributes['extra']['cowc'], $this->longCountry->getCowc());
     }
 
-    /** @test */
-    public function it_returns_null_when_missing_cowc()
+    #[Test]
+    public function it_returns_null_when_missing_cowc(): void
     {
         $this->longCountry->setAttributes([]);
 
         $this->assertNull($this->longCountry->getCowc());
     }
 
-    /** @test */
-    public function it_returns_cown()
+    #[Test]
+    public function it_returns_cown(): void
     {
         $this->assertEquals($this->longAttributes['extra']['cown'], $this->longCountry->getCown());
     }
 
-    /** @test */
-    public function it_returns_null_when_missing_cown()
+    #[Test]
+    public function it_returns_null_when_missing_cown(): void
     {
         $this->longCountry->setAttributes([]);
 
         $this->assertNull($this->longCountry->getCown());
     }
 
-    /** @test */
-    public function it_returns_fao()
+    #[Test]
+    public function it_returns_fao(): void
     {
         $this->assertEquals($this->longAttributes['extra']['fao'], $this->longCountry->getFao());
     }
 
-    /** @test */
-    public function it_returns_null_when_missing_fao()
+    #[Test]
+    public function it_returns_null_when_missing_fao(): void
     {
         $this->longCountry->setAttributes([]);
 
         $this->assertNull($this->longCountry->getFao());
     }
 
-    /** @test */
-    public function it_returns_imf()
+    #[Test]
+    public function it_returns_imf(): void
     {
         $this->assertEquals($this->longAttributes['extra']['imf'], $this->longCountry->getImf());
     }
 
-    /** @test */
-    public function it_returns_null_when_missing_imf()
+    #[Test]
+    public function it_returns_null_when_missing_imf(): void
     {
         $this->longCountry->setAttributes([]);
 
         $this->assertNull($this->longCountry->getImf());
     }
 
-    /** @test */
-    public function it_returns_ar5()
+    #[Test]
+    public function it_returns_ar5(): void
     {
         $this->assertEquals($this->longAttributes['extra']['ar5'], $this->longCountry->getAr5());
     }
 
-    /** @test */
-    public function it_returns_null_when_missing_ar5()
+    #[Test]
+    public function it_returns_null_when_missing_ar5(): void
     {
         $this->longCountry->setAttributes([]);
 
         $this->assertNull($this->longCountry->getAr5());
     }
 
-    /** @test */
-    public function it_returns_address_format()
+    #[Test]
+    public function it_returns_address_format(): void
     {
         $this->assertEquals($this->longAttributes['extra']['address_format'], $this->longCountry->getAddressFormat());
     }
 
-    /** @test */
-    public function it_returns_null_when_missing_address_format()
+    #[Test]
+    public function it_returns_null_when_missing_address_format(): void
     {
         $this->longCountry->setAttributes([]);
 
         $this->assertNull($this->longCountry->getAddressFormat());
     }
 
-    /** @test */
-    public function it_returns_whether_eu_member()
+    #[Test]
+    public function it_returns_whether_eu_member(): void
     {
         $this->assertEquals($this->longAttributes['extra']['eu_member'], $this->longCountry->isEuMember());
     }
 
-    /** @test */
-    public function it_returns_null_when_missing_eu_member_status()
+    #[Test]
+    public function it_returns_null_when_missing_eu_member_status(): void
     {
         $this->longCountry->setAttributes([]);
 
         $this->assertNull($this->longCountry->isEuMember());
     }
 
-    /** @test */
-    public function it_returns_whether_data_protection()
+    #[Test]
+    public function it_returns_whether_data_protection(): void
     {
         $this->assertEquals($this->longAttributes['extra']['data_protection'], $this->longCountry->getDataProtection());
     }
 
-    /** @test */
-    public function it_returns_null_when_missing_data_protection_status()
+    #[Test]
+    public function it_returns_null_when_missing_data_protection_status(): void
     {
         $this->longCountry->setAttributes([]);
 
         $this->assertNull($this->longCountry->getDataProtection());
     }
 
-    /** @test */
-    public function it_returns_vat_rates()
+    #[Test]
+    public function it_returns_vat_rates(): void
     {
         $this->assertEquals($this->longAttributes['extra']['vat_rates'], $this->longCountry->getVatRates());
     }
 
-    /** @test */
-    public function it_returns_null_when_missing_vat_rates()
+    #[Test]
+    public function it_returns_null_when_missing_vat_rates(): void
     {
         $this->longCountry->setAttributes([]);
 
         $this->assertNull($this->longCountry->getVatRates());
     }
 
-    /** @test */
-    public function it_returns_emoji_from_longlist()
+    #[Test]
+    public function it_returns_emoji_from_longlist(): void
     {
         $this->assertEquals($this->longAttributes['extra']['emoji'], $this->longCountry->getEmoji());
     }
 
-    /** @test */
-    public function it_returns_emoji_from_shortlist()
+    #[Test]
+    public function it_returns_emoji_from_shortlist(): void
     {
         $this->assertEquals($this->shortAttributes['emoji'], $this->shortCountry->getEmoji());
     }
 
-    /** @test */
-    public function it_returns_null_when_missing_emoji()
+    #[Test]
+    public function it_returns_null_when_missing_emoji(): void
     {
         $this->longCountry->setAttributes([]);
 
         $this->assertNull($this->longCountry->getEmoji());
     }
 
-    /** @test */
-    public function it_returns_geojson()
+    #[Test]
+    public function it_returns_geojson(): void
     {
-        $file = __DIR__.'/../../resources/geodata/'.mb_strtolower($this->longCountry->getIsoAlpha2()).'.json';
+        $file = __DIR__.'/../../resources/geodata/'.mb_strtolower((string) $this->longCountry->getIsoAlpha2()).'.json';
 
         $this->assertEquals(file_get_contents($file), $this->longCountry->getGeoJson());
     }
 
-    /** @test */
-    public function it_returns_null_when_missing_geojson()
+    #[Test]
+    public function it_returns_null_when_missing_geojson(): void
     {
         $this->longCountry->setAttributes([]);
 
         $this->assertNull($this->longCountry->getGeoJson());
     }
 
-    /** @test */
-    public function it_returns_flag()
+    #[Test]
+    public function it_returns_flag(): void
     {
-        $file = __DIR__.'/../../resources/flags/'.mb_strtolower($this->longCountry->getIsoAlpha2()).'.svg';
+        $file = __DIR__.'/../../resources/flags/'.mb_strtolower((string) $this->longCountry->getIsoAlpha2()).'.svg';
 
         $this->assertEquals(file_get_contents($file), $this->longCountry->getFlag());
     }
 
-    /** @test */
-    public function it_returns_null_when_missing_flag()
+    #[Test]
+    public function it_returns_null_when_missing_flag(): void
     {
         $this->longCountry->setAttributes([]);
 
         $this->assertNull($this->longCountry->getFlag());
     }
 
-    /** @test */
-    public function it_returns_divisions()
+    #[Test]
+    public function it_returns_divisions(): void
     {
-        $file = __DIR__.'/../../resources/divisions/'.mb_strtolower($this->longCountry->getIsoAlpha2()).'.json';
+        $file = __DIR__.'/../../resources/divisions/'.mb_strtolower((string) $this->longCountry->getIsoAlpha2()).'.json';
 
         $this->assertEquals(json_decode(file_get_contents($file), true), $this->longCountry->getDivisions());
     }
 
-    /** @test */
-    public function it_returns_null_when_missing_divisions()
+    #[Test]
+    public function it_returns_null_when_missing_divisions(): void
     {
         $this->longCountry->setAttributes([]);
 
         $this->assertNull($this->longCountry->getDivisions());
     }
 
-    /** @test */
-    public function it_returns_division()
+    #[Test]
+    public function it_returns_division(): void
     {
         $this->assertEquals($this->longAttributes['divisions']['ALX'], $this->longCountry->getDivision('ALX'));
     }
 
-    /** @test */
-    public function it_returns_null_when_missing_division()
+    #[Test]
+    public function it_returns_null_when_missing_division(): void
     {
         $this->longCountry->setAttributes([]);
 
         $this->assertNull($this->longCountry->getDivisions());
     }
 
-    /** @test */
-    public function it_returns_timezones()
+    #[Test]
+    public function it_returns_timezones(): void
     {
         $this->assertEquals(['Africa/Cairo'], $this->shortCountry->getTimezones());
     }
 
-    /** @test */
-    public function it_returns_null_when_missing_timezones()
+    #[Test]
+    public function it_returns_null_when_missing_timezones(): void
     {
         $this->shortCountry->setAttributes([]);
 
         $this->assertNull($this->shortCountry->getTimezones());
     }
 
-    /** @test */
-    public function it_returns_locales()
+    #[Test]
+    public function it_returns_locales(): void
     {
         $this->assertEquals(['ar_EG'], $this->shortCountry->getLocales());
     }
 
-    /** @test */
-    public function it_returns_null_when_missing_locales()
+    #[Test]
+    public function it_returns_null_when_missing_locales(): void
     {
         $this->shortCountry->setAttributes([]);
 
