@@ -6,11 +6,12 @@ namespace Rinvex\Country\Tests\Unit;
 
 use Rinvex\Country\Country;
 use PHPUnit\Framework\TestCase;
+use PHPUnit\Framework\Attributes\Test;
 
 class HelpersTest extends TestCase
 {
-    /** @test */
-    public function it_returns_country_data()
+    #[Test]
+    public function it_returns_country_data(): void
     {
         $egypt = [
             'name' => [
@@ -115,8 +116,8 @@ class HelpersTest extends TestCase
         $this->assertEquals(new Country($egypt), country('eg'));
     }
 
-    /** @test */
-    public function it_returns_country_array_shortlist()
+    #[Test]
+    public function it_returns_country_array_shortlist(): void
     {
         $this->assertEquals(250, count(countries()));
         $this->assertIsArray(countries()['eg']);
@@ -124,8 +125,8 @@ class HelpersTest extends TestCase
         $this->assertArrayNotHasKey('geo', countries()['eg']);
     }
 
-    /** @test */
-    public function it_returns_courrencies_longlist()
+    #[Test]
+    public function it_returns_courrencies_longlist(): void
     {
         $this->assertEquals(165, count(currencies(true)));
         $this->assertArrayHasKey('EGP', currencies());
@@ -136,8 +137,8 @@ class HelpersTest extends TestCase
         $this->assertEquals('2', currencies(true)['EGP']['iso_4217_minor_unit']);
     }
 
-    /** @test */
-    public function it_returns_courrencies_shortlist()
+    #[Test]
+    public function it_returns_courrencies_shortlist(): void
     {
         $this->assertEquals(165, count(currencies()));
         $this->assertArrayHasKey('EGP', currencies());
