@@ -10,10 +10,8 @@ class CountryLoaderException extends Exception
 {
     /**
      * Create a new exception instance.
-     *
-     * @return static
      */
-    public static function invalidCountry()
+    public static function invalidCountry(): static
     {
         return new static('Country code may be misspelled, invalid, or data not found on server!');
     }
