@@ -2,6 +2,7 @@
 
 declare(strict_types=1);
 
+use Rinvex\Country\Country;
 use Rinvex\Country\CountryLoader;
 use Rinvex\Country\CurrencyLoader;
 
@@ -12,7 +13,7 @@ if (! function_exists('country')) {
      * @param string $code
      * @param bool   $hydrate
      *
-     * @return \Rinvex\Country\Country|array
+     * @return Country|array
      */
     function country($code, $hydrate = true)
     {
@@ -41,10 +42,8 @@ if (! function_exists('currencies')) {
      *
      * @param bool $longlist
      * @param bool $hydrate
-     *
-     * @return array
      */
-    function currencies($longlist = false)
+    function currencies($longlist = false): array
     {
         return CurrencyLoader::currencies($longlist);
     }
